juggler: use non-pooled connections for long-lived pubsub, results and calls polling
Closing a connection that had subscribed to a channel had a race
as a pooled connection doesn`t get closed, only released back
to the pool. This doesn`t work in the case of long-polling
goroutines (pubsub, results and calls) since it needs a closed
connection to detect that it`s time to exit.

This fixes it by adding the Dial function field on the Broker,
and using it to create non-pooled redis connections for those
cases.